requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
alpaca-py==0.43.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
arch==8.0.0
astropy==7.2.0
astropy-iers-data==0.2026.3.9.0.47.45
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
bidict==0.23.1
billiard==4.2.4
blinker==1.9.0
bt==1.1.4
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.5
clarabel==0.11.1
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.3
cryptography==43.0.3
curl_cffi==0.14.0
cvxpy==1.8.1
cycler==0.12.1
dateparser==1.3.0
decorator==5.2.1
dill==0.4.1
dnspython==2.8.0
eventlet==0.40.4
executing==2.2.1
factory_boy==3.3.3
Faker==40.11.0
fakeredis==2.34.1
ffn==1.1.4
filelock==3.25.2
finnhub-python==2.4.27
Flask==3.1.3
Flask-Admin==2.0.2
Flask-Login==0.6.3
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-Talisman==1.1.0
Flask-WTF==1.2.2
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
gunicorn==23.0.0
h11==0.16.0
highspy==1.13.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.3
iniconfig==2.3.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
kombu==5.6.2
llvmlite==0.44.0
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.18.0
networkx==3.6.1
numba==0.61.2
numpy==2.2.6
osqp==1.1.1
packaging==26.0
pandas==2.3.3
pandas-datareader==0.10.0
pandas-ta==0.4.71b0
parso==0.8.6
patsy==1.0.2
peewee==4.0.1
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
polygon-api-client==1.16.3
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.0
psycopg2-binary==2.9.11
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==3.0.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyerfa==2.0.1.5
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
pyportfolioopt==1.6.0
PyPrind==2.11.3
pytest==9.0.2
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-socketio==5.16.1
pytz==2026.1.post1
PyYAML==6.0.3
quantstats==0.0.81
redis==5.3.1
regex==2026.2.28
requests==2.32.5
Riskfolio-Lib==6.1.1
schedule==1.2.2
scikit-base==0.13.1
scikit-learn==1.8.0
scipy==1.17.1
scs==3.2.11
seaborn==0.13.2
setuptools==82.0.1
simple-websocket==1.1.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
sseclient-py==1.9.0
stack-data==0.6.3
statsmodels==0.14.6
structlog==25.5.0
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.6.3
vaderSentiment==3.3.2
vectorbt==0.28.4
vine==5.1.0
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
wsproto==1.3.2
WTForms==3.2.1
xlsxwriter==3.2.9
yarl==1.23.0
yfinance==0.2.66
